Fix CSRF bug on authentication
Since we had registered omniauth twice (once with devise and once via the Omniauth initializer), the middleware was running twice on each callback request. This was causing some issues where credentials wouldn't be assigned after sign in.

The unfortunate side effect of this PR is that a user will need to add additional callback URLs in their Google Oauth configuration. Have spent some time looking at this but can't find a way around at the moment.